requirements.txt
gradient
python-dotenv
fastapi
uvicorn[standard]
httpx
pydantic
